# Dramatically improve watch mode performance.