--- a/packages/jest-config/src/__tests__/getMaxWorkers.test.ts
+++ b/packages/jest-config/src/__tests__/getMaxWorkers.test.ts
@@ -32,7 +32,7 @@ describe('getMaxWorkers', () => {
 
   it('Returns based on the number of cpus', () => {
     expect(getMaxWorkers({})).toBe(3);
-    expect(getMaxWorkers({watch: true})).toBe(2);
+    expect(getMaxWorkers({watch: true})).toBe(3);
   });
 
   describe('% based', () => {