--- a/packages/jest-haste-map/src/ModuleMap.ts
+++ b/packages/jest-haste-map/src/ModuleMap.ts
@@ -32,11 +32,11 @@ export type SerializableModuleMap = {
 };
 
 export default class ModuleMap {
+  static DuplicateHasteCandidatesError: typeof DuplicateHasteCandidatesError;
+  private static nextUniqueID = 1; // Must be truthy.
   public readonly uniqueID: number;
   private readonly _raw: RawModuleMap;
   private json: SerializableModuleMap | undefined;
-  static DuplicateHasteCandidatesError: typeof DuplicateHasteCandidatesError;
-  private static nextUniqueID = 1; // Must be truthy.
 
   constructor(raw: RawModuleMap) {
     this.uniqueID = ModuleMap.nextUniqueID;