--- a/packages/jest-runner/src/__tests__/testRunner.test.js
+++ b/packages/jest-runner/src/__tests__/testRunner.test.js
@@ -51,7 +51,6 @@ test('injects the serializable module map into each worker in watch mode', () =>
             context: runContext,
             globalConfig,
             path: './file.test.js',
-            serializableModuleMap,
           },
         ],
         [
@@ -60,7 +59,6 @@ test('injects the serializable module map into each worker in watch mode', () =>
             context: runContext,
             globalConfig,
             path: './file2.test.js',
-            serializableModuleMap,
           },
         ],
       ]);
@@ -89,9 +87,7 @@ test('does not inject the serializable module map in serial mode', () => {
             config,
             context: runContext,
             globalConfig,
-            moduleMapUniqueID: null,
             path: './file.test.js',
-            serializableModuleMap: null,
           },
         ],
         [
@@ -99,9 +95,7 @@ test('does not inject the serializable module map in serial mode', () => {
             config,
             context: runContext,
             globalConfig,
-            moduleMapUniqueID: null,
             path: './file2.test.js',
-            serializableModuleMap: null,
           },
         ],
       ]);