--- a/packages/jest-runner/src/index.ts
+++ b/packages/jest-runner/src/index.ts
@@ -11,7 +11,7 @@ import exit from 'exit';
 import throat from 'throat';
 import Worker from 'jest-worker';
 import runTest from './runTest';
-import {worker} from './testWorker';
+import {worker, SerializableResolver} from './testWorker';
 import {
   OnTestFailure,
   OnTestStart,
@@ -103,11 +103,31 @@ class TestRunner {
     onResult: OnTestSuccess,
     onFailure: OnTestFailure,
   ) {
+    let resolvers: Map<string, SerializableResolver> | undefined = undefined;
+    if (watcher.isWatchMode()) {
+      resolvers = new Map();
+      for (const test of tests) {
+        if (!resolvers.has(test.context.config.name)) {
+          resolvers.set(test.context.config.name, {
+            config: test.context.config,
+            serializableModuleMap: test.context.moduleMap.toJSON(),
+          });
+        }
+      }
+    }
+
     const worker = new Worker(TEST_WORKER_PATH, {
       exposedMethods: ['worker'],
       forkOptions: {stdio: 'pipe'},
       maxRetries: 3,
       numWorkers: this._globalConfig.maxWorkers,
+      setupArgs: resolvers
+        ? [
+            {
+              serializableResolvers: Array.from(resolvers.values()),
+            },
+          ]
+        : undefined,
     }) as WorkerInterface;
 
     if (worker.getStdout()) worker.getStdout().pipe(process.stdout);
@@ -134,13 +154,7 @@ class TestRunner {
               Array.from(this._context.changedFiles),
           },
           globalConfig: this._globalConfig,
-          moduleMapUniqueID: watcher.isWatchMode()
-            ? test.context.moduleMap.uniqueID
-            : null,
           path: test.path,
-          serializableModuleMap: watcher.isWatchMode()
-            ? test.context.moduleMap.toJSON()
-            : null,
         });
       });