## Summary

Resolves #7341 

This PR dramatically improves watch mode performance, bringing it in line with single run mode performance. It accomplishes that by:
- Workers previously initialized a new `ModuleMap` and `Resolver` for every test in watch mode. Now, those objects are only initialized once when the worker is setup.
- In the main thread, caching the conversion of `ModuleMap` to a JSON-friendly object.
- Allowing watch mode to use the same number of CPUs as single run mode.

## Benchmarks

I benchmarked against Jest's own test suite, excluding e2e tests which don't provide good signal because they individually take a long time (so startup time for the test is marginalized). The numbers show that running in Watch mode previously added an extra 35%~ of runtime to the tests but that has now been reduced to almost nothing.

Watch mode should now just be paying a one-time initial cost for each worker when the haste map changes instead of paying that same cost for _every_ test run.

### branch: master

`yarn jest ./packages`
Run time: 15.091s

`yarn jest ./packages --watch`
Run time: 23.234s

### branch: watch-performance

`yarn jest ./packages`
Run time: 14.973s

`yarn jest ./packages --watch`
Run time: 15.196s


## Test plan

- All tests pass.
- Benchmarked to verify the performance wins.
- Verified that when the haste map is updated, the update is propagated out to all workers.