@SimenB 2019-03-25
Shouldn't that work the same regardless of watch mode? Why should watch mode use less cores than a normal run?