@cpojer 2019-03-25
Mainly because watch mode is something you have open as a service, often in the background, and it runs automatically while writing/saving code. When you invoke `jest` on the other hand, the user intentionally invokes the test runner knowingly and is more likely to be ok with the resource usage.