@SimenB 2019-03-25
Right, makes sense