@scotthovestadt 2019-03-25
@cpojer I actually checked the blame but it was lost to an earlier refactor. :)

It was only a small part of this PR anyway, if anything I thought it might have been a mistake. I’ll submit a PR to revert that piece.