@SimenB 2019-03-25
Can that PR include doc changes which explains what these values are and why? That was part of the confusion in #7341