@scotthovestadt 2019-03-25
Sure, will do.

I just validated (as is pretty obvious) that even without the CPU change this is a massive performance boost. At 300k files in the map, I’m seeing things run almost twice as fast.

I’ll work on reverting that piece with documentation today.